+++ src/proxies/withAPIGateway.ts
@@ -5,7 +5,7 @@ import {
   APIGatewayProxyEventQueryStringParameters,
   Context,
 } from 'aws-lambda';
-import { parseEvent } from '../utils/APIGatewayRequest';
+import {HolpHeaders, parseEvent} from '../utils/APIGatewayRequest';
 import {
   Logger,
   APIGatewayResponse,
@@ -14,8 +14,9 @@ import {
 } from '../utils/APIGatewayResponse';
 
 export interface APIGatewayParsedEvent
-  extends Omit<APIGatewayProxyEvent, 'body'> {
+  extends Omit<APIGatewayProxyEvent, 'body' | 'headers'> {
   body: unknown;
+  headers: HolpHeaders;
   queryStringParameters: APIGatewayProxyEventQueryStringParameters;
   pathParameters: APIGatewayProxyEventPathParameters;
   rawHeaders: APIGatewayProxyEventHeaders;
@@ -23,13 +24,24 @@ export interface APIGatewayParsedEvent
 
 type Handler<TEvent> = (
   event: TEvent,
-  context: Context
+  context: Context,
 ) => Promise<APIGatewayResponse>;
 
+interface CorsSettings {
+  origin?: string;
+  allowCredentials?: boolean;
+}
 export interface withAPIGatewayOptions {
   logger?: Logger;
-  cors?: boolean;
+  cors?: CorsSettings;
 }
+
+const defaultLogger: Logger = {
+  debug: console.log,
+  info: console.log,
+  warn: console.log,
+  error: console.log,
+};
 
 /**
  * Higher Order Function that wraps your lambda handler code running
@@ -38,17 +50,19 @@ export interface withAPIGatewayOptions {
  *
  * @param handler - Your lambda handler function
  * @param options - Options for the wrapper
- * @param options.logger - A logger instance
- * @param options.cors - Whether to add CORS headers to the response
+ * @param options.logger - A logger instance (console.log by default)
+ * @param options.cors - CORS configuration.
  * @returns A lambda handler function
  */
 export function withAPIGateway<TEvent extends APIGatewayParsedEvent>(
   handler: Handler<TEvent>,
-  options: withAPIGatewayOptions = {}
+  options: withAPIGatewayOptions = {logger: defaultLogger},
 ) {
   return async function (event: APIGatewayProxyEvent, context: Context) {
     try {
-      options.logger?.debug(event, 'START');
+      if (process.env.DEBUG === 'true') {
+        options.logger?.debug({...event, msg: '[INCOMING_EVENT]'});
+      }
       const parsedEvent = parseEvent<TEvent>(event);
       const response = await handler(parsedEvent, context);
       return responseFactory(response, options);